requirements.txt
selenium~=4.0.0
lxml~=4.6.4
aiohttp~=3.8.0
requests~=2.26.0
python-dotenv~=0.19.1